fix: parse f32 before converting to u32

+++ box_it/src/lib.rs
@@ -3,7 +3,7 @@ pub fn transform_and_save_on_heap(s: String) -> Box<Vec<u32>> {
         s.split_whitespace()
             .map(|s| {
                 if let Some(stripped) = s.strip_suffix(|c| c == 'k' || c == 'K') {
-                    stripped.parse::<u32>().expect("invalid number") * 1000
+                    (stripped.parse::<f32>().expect("invalid number") * 1000.0) as u32 
                 } else {
                     s.parse::<u32>().expect("Invalid number")
                 }
